fixup! treewide: make type casts explicit

+++ crypto.c
@@ -311,12 +311,12 @@ dtls_ccm_encrypt(aes128_ccm_t *ccm_ctx, const unsigned char *src, size_t srclen,
   return len;
 }
 
-static size_t
+static ssize_t
 dtls_ccm_decrypt(aes128_ccm_t *ccm_ctx, const unsigned char *src,
 		 size_t srclen, unsigned char *buf,
 		 const unsigned char *nonce,
 		 const unsigned char *aad, size_t la) {
-  long int len;
+  ssize_t len;
   (void)src;
 
   assert(ccm_ctx);
